Move wish list tests to DetailedItemFragmentTest
